[TASK] #84610 | Extract information collection into service
