+++ Classes/Service/SystemInformationInterface.php
@@ -6,4 +6,10 @@ namespace Psychomieze\AdminpanelExtended\Service;
 /**
  * Interface SystemInformationInterface
  */
-interface SystemInformationInterface {}
+interface SystemInformationInterface
+{
+    /**
+     * Collect the system information
+     */
+    public function collectInformation(): array;
+}
